+++ app/upgrades.go
@@ -14,7 +14,6 @@ func (app *App) RegisterUpgradeHandlers(cdc codec.BinaryCodec, maxVersion int64)
 	planName := "v1.8"
 	app.UpgradeKeeper.SetUpgradeHandler(planName,
 		func(ctx context.Context, _ upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
-			app.ActivateCRC21PrecompileBlocks()
 			return app.ModuleManager.RunMigrations(ctx, app.configurator, fromVM)
 		},
 	)